Quick refactor to tidy code
* Remove unneeded `[Serializable]` attribute tags.
* Remove empty folder `Writers`
* Refactor test only extension in main project to the method in consuming test.
