+++ EndpointForge.Abstractions/EndpointForge.Abstractions.csproj
@@ -8,9 +8,5 @@
     <ItemGroup>
         <FrameworkReference Include="Microsoft.AspNetCore.App" />
     </ItemGroup>
-
-    <ItemGroup>
-      <Folder Include="Writers\" />
-    </ItemGroup>
     
 </Project>
