+++ EndpointForge.Abstractions/Models/EndpointForgeParameterDetails.cs
@@ -1,4 +1,3 @@
 namespace EndpointForge.Abstractions.Models;
 
-[Serializable]
 public record EndpointForgeParameterDetails(string Identifier, string Value);
