+++ EndpointForge.Abstractions/Models/EndpointRoutingDetails.cs
@@ -1,4 +1,3 @@
 namespace EndpointForge.Abstractions.Models;
 
-[Serializable]
 public record EndpointRoutingDetails(string Route, string Method);
