+++ EndpointForge.WebApi/Extensions/HttpRequestExtensions.cs
@@ -1,4 +1,3 @@
-using System.Text;
 using System.Text.Json;
 using EndpointForge.Abstractions.Exceptions;
 
@@ -6,18 +5,6 @@ namespace EndpointForge.WebApi.Extensions;
 
 public static class HttpRequestExtensions
 {
-    public static async Task WriteAsJsonAsync<T>(this HttpRequest request, T data)
-    {
-        request.ContentType = "application/json";
-        
-        var json = JsonSerializer.Serialize(data);
-        var bytes = Encoding.UTF8.GetBytes(json);
-
-        request.Body = new MemoryStream(bytes);
-        await request.Body.FlushAsync();
-        request.Body.Seek(0, SeekOrigin.Begin);
-    }
-
     public static async Task<T> TryDeserializeRequestAsync<T>(
         this HttpRequest request) where T : class
     {
